Nav bar restructure
