+++ jams/routes/frontend/private/admin.py
@@ -2,7 +2,6 @@
 from flask import Blueprint, render_template
 from flask_security import login_required
 from jams.decorators import role_based_access_control_fe
-from jams.models import Event
 
 url_prefix='/private/admin'
 
@@ -36,26 +35,6 @@ def settings_roles():
 def settings_auth_sources():
     return render_template(f'{url_prefix}/settings/settings-auth_sources.html')
 
-# Events
-@bp.route('/events')
-@login_required
-@role_based_access_control_fe
-def events():
-    return render_template(f'{url_prefix}/events/events.html')
-
-@bp.route('/events/add')
-@login_required
-@role_based_access_control_fe
-def add_event():
-    return render_template(f'{url_prefix}/events/add_event.html')
-
-@bp.route('/events/<int:event_id>/edit')
-@login_required
-@role_based_access_control_fe
-def edit_event(event_id):
-    Event.query.filter_by(id=event_id).first_or_404() # Make sure the event exists before rendering the template
-    return render_template(f'{url_prefix}/events/edit_event.html')
-
 ## Other
 
 @bp.route('/user_management')
@@ -63,9 +42,3 @@ def edit_event(event_id):
 @role_based_access_control_fe
 def user_management():
     return render_template(f'{url_prefix}/user_management.html')
-
-@bp.route('/schedule_planner')
-@login_required
-@role_based_access_control_fe
-def schedule_planner():
-    return render_template(f'{url_prefix}/schedule_planner.html')
