+++ jams/static/ts/management/events/add_event.ts
