+++ jams/static/ts/management/events/edit_event.ts
