+++ jams/templates/private/event/schedule_planner.html
@@ -141,7 +141,7 @@
     </div>
 </div>
 
-<script src="{{ url_for('static', filename='js/admin/schedule_planner.js') }}" type="module" defer></script>
+<script src="{{ url_for('static', filename='js/event/schedule_planner.js') }}" type="module" defer></script>
 <link rel="stylesheet" href="{{ url_for('static', filename='css/admin/schedule_planner.css') }}">
 <link rel="stylesheet" href="{{ url_for('static', filename='css/general/schedule_grid.css') }}">
 <link rel="stylesheet" href="{{ url_for('static', filename='css/general/workshop_card.css') }}">
