+++ jams/templates/private/nav.html
@@ -73,7 +73,7 @@
         <!-- Other Navigation Items -->
         {% if current_user.is_authenticated %}
         <!--Event-->
-        {% if user_has_access_to_page('fire_list', 'attendee_list') %}
+        {% if user_has_access_to_page('fire_list', 'attendee_list', 'schedule_planner') %}
         <li class="nav-item dropdown">
           <a class="nav-link dropdown-toggle" href="#" data-bs-toggle="dropdown" data-bs-auto-close="outside"
             role="button" aria-expanded="false">
@@ -82,6 +82,11 @@
             </span>
           </a>
           <div class="dropdown-menu">
+            {% if user_has_access_to_page('schedule_planner') %}
+            <a class="dropdown-item" href="{{ url_for('routes.frontend.private.event.schedule_planner')}}">
+              Schedule Planner
+            </a>
+            {% endif %}
             {% if user_has_access_to_page('fire_list') %}
             <a class="dropdown-item" href="{{ url_for('routes.frontend.private.event.fire_list')}}">
               Fire List
@@ -122,7 +127,7 @@
         </li>
         {% endif %}
         <!--Management-->
-        {% if user_has_access_to_page('workshops', 'locations_timeslots') %}
+        {% if user_has_access_to_page('workshops', 'locations_timeslots', 'events') %}
         <li class="nav-item dropdown">
           <a class="nav-link dropdown-toggle" href="#" data-bs-toggle="dropdown" data-bs-auto-close="outside"
             role="button" aria-expanded="false">
@@ -131,6 +136,11 @@
             </span>
           </a>
           <div class="dropdown-menu">
+            {% if user_has_access_to_page('events') %}
+            <a class="dropdown-item" href="{{ url_for('routes.frontend.private.management.events')}}">
+              Events
+            </a>
+            {% endif %}
             {% if user_has_access_to_page('workshops') %}
             <a class="dropdown-item" href="{{ url_for('routes.frontend.private.management.workshops')}}">
               Workshops
@@ -145,7 +155,7 @@
         </li>
         {% endif %}
         <!--Admin-->
-        {% if user_has_access_to_page('settings', 'user_management', 'events', 'schedule_planner') %}
+        {% if user_has_access_to_page('settings', 'user_management') %}
         <li class="nav-item dropdown">
           <a class="nav-link dropdown-toggle" href="#" data-bs-toggle="dropdown" data-bs-auto-close="outside"
             role="button" aria-expanded="false">
@@ -162,16 +172,6 @@
             {% if user_has_access_to_page('user_management') %}
             <a class="dropdown-item" href="{{ url_for('routes.frontend.private.admin.user_management')}}">
               User Management
-            </a>
-            {% endif %}
-            {% if user_has_access_to_page('events') %}
-            <a class="dropdown-item" href="{{ url_for('routes.frontend.private.admin.events')}}">
-              Events
-            </a>
-            {% endif %}
-            {% if user_has_access_to_page('schedule_planner') %}
-            <a class="dropdown-item" href="{{ url_for('routes.frontend.private.admin.schedule_planner')}}">
-              Schedule Planner
             </a>
             {% endif %}
           </div>
